% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eponge-package.R
\docType{package}
\name{eponge-package}
\alias{eponge-package}
\alias{eponge}
\title{Package eponge}
\description{
Provides a set of functions, which facilitates removing objects from an environment.
It allows to delete objects specified with regular expression or
with other conditions (e.g. if object is numeric), using one function call.
}
\author{
Krzysztof Joachimiak
}
\keyword{package}
