% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erase-by-type.R
\name{erase_data}
\alias{erase_data}
\title{Remove all objects, which are listed in `Data` section in RStudio}
\usage{
erase_data(pattern = NULL, envir = globalenv(), verbose = TRUE)
}
\arguments{
\item{pattern}{regex pattern to select a set of functions; default: NULL}

\item{envir}{environment; default: global environment}

\item{verbose}{print removed objects' names}
}
\description{
Remove all objects, which are listed in `Data` section in RStudio
}
\examples{
cars.2 <- cars
test_fun <- function(x) x + 2
value <- 7
erase_non_functions(verbose = TRUE)
}
