% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erase-by-type.R
\name{erase_df}
\alias{erase_df}
\title{Remove all the `data.frame` objects}
\usage{
erase_df(pattern = NULL, envir = globalenv(), verbose = FALSE)
}
\arguments{
\item{pattern}{regex pattern to select a set of functions; default: NULL}

\item{envir}{environment; default: global environment}

\item{verbose}{print removed objects' names}
}
\description{
Remove all the `data.frame` objects
}
\examples{
cars.2 <- cars
test_fun <- function(x) x + 2
value <- 7
erase_non_functions(verbose = TRUE)
}
