% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erase-masking.R
\name{erase_masking}
\alias{erase_masking}
\alias{erase_masking_functions}
\title{Erase objects from GlobalEnv, which are masking objects from attached packages}
\usage{
erase_masking(pattern = NULL, verbose = FALSE)

erase_masking_functions(pattern = NULL, verbose = FALSE)
}
\arguments{
\item{pattern}{a regex pattern}

\item{verbose}{print removed objects' names}
}
\description{
Erase objects from GlobalEnv, which are masking objects from attached packages
}
\details{
We have to highglight, that for now it only allows us to remove objects from
the Global Environment
}
\examples{
assign('matrix', matrix(0, 3, 3), envir = globalenv())
assign('gamma', 0.9, envir = globalenv())
erase_masking(verbose = TRUE)
ls(envir = globalenv())
}
