% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erase-masking.R
\name{erase_masking_objects}
\alias{erase_masking_objects}
\alias{erase_masking_functions}
\title{Erase objects from GlobalEnv, which are masking objects from attached packages}
\usage{
erase_masking_objects(pattern = NULL, verbose = FALSE)

erase_masking_functions(pattern = NULL, verbose = FALSE)
}
\arguments{
\item{pattern}{a regex pattern}

\item{verbose}{print removed objects' names}
}
\description{
Erase objects from GlobalEnv, which are masking objects from attached packages
}
\details{
We have to highglight, that for now it only allows us to remove objects from
the Global Environment
}
\examples{
matrix <- matrix(0, 3, 3)
gamma <- 0.9
erase_masking_objects(verbose = TRUE)
}
