% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erase-by-type.R
\name{erase_non_functions}
\alias{erase_non_functions}
\title{Remove all the objects, that are not functions}
\usage{
erase_non_functions(pattern = NULL, envir = globalenv(),
  verbose = TRUE)
}
\arguments{
\item{pattern}{regex pattern to select a set of functions; default: NULL}

\item{envir}{environment; default: global environment}
}
\description{
Remove all the objects, that are not functions
}
\examples{
cars.2 <- cars
test_fun <- function(x) x + 2
value <- 7
erase_non_functions(verbose = TRUE)
}
