% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erase-by-type.R
\name{erase_values}
\alias{erase_values}
\title{Remove all objects, which are listed in `Values` section in RStudio}
\usage{
erase_values(pattern = NULL, envir = globalenv(), verbose = FALSE)
}
\arguments{
\item{pattern}{regex pattern to select a set of objects; default: NULL}

\item{envir}{environment; default: global environment}

\item{verbose}{print removed objects' names}
}
\description{
Remove all objects, which are listed in `Values` section in RStudio
}
\examples{
cars.2 <- cars
test_fun <- function(x) x + 2
value <- 7
erase_non_functions(envir = environment(), verbose = TRUE)
}
