% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erase-by-type.R
\name{erase_functions}
\alias{erase_functions}
\title{Remove (all) functions from environment}
\usage{
erase_functions(pattern = NULL, envir = parent.frame(),
  verbose = FALSE)
}
\arguments{
\item{pattern}{regex pattern to select a set of objects; default: NULL}

\item{envir}{environment; default: caller environment}

\item{verbose}{print removed objects' names}
}
\value{
NULL (function returns nothing)
}
\description{
Remove (all) functions from environment
}
\details{
Function can be used with envir = globalenv() argument.
Be careful: uncontrolled use may cause undesired side effects.
}
\examples{
create_data <- function() data.frame(a = 1:10, b = 11:20)
x <- cars
y <- 1:20
z <- function(x) x +2
erase_functions()
ls()
}
